+++ Binance-Spot-API/Model/Market/TickerPrice.cs
@@ -8,31 +8,31 @@ namespace Binance_Spot_API.Model.Market
         [JsonProperty("symbol")]
         public string Symbol { get; set; }
         [JsonProperty("priceChange")]
-        public double PriceChange { get; set; }
+        public decimal PriceChange { get; set; }
         [JsonProperty("priceChangePercent")]
-        public double PriceChangePercent { get; set; }
+        public decimal PriceChangePercent { get; set; }
         [JsonProperty("weightedAvgPrice")]
-        public double WeightedAvgPrice { get; set; }
+        public decimal WeightedAvgPrice { get; set; }
         [JsonProperty("prevClosePrice")]
-        public double PrevClosePrice { get; set; }
+        public decimal PrevClosePrice { get; set; }
         [JsonProperty("lastPrice")]
-        public double LastPrice { get; set; }
+        public decimal LastPrice { get; set; }
         [JsonProperty("lastQty")]
-        public double LastQuantity { get; set; }
+        public decimal LastQuantity { get; set; }
         [JsonProperty("bidPrice")]
-        public double BidPrice { get; set; }
+        public decimal BidPrice { get; set; }
         [JsonProperty("askPrice")]
-        public double AskPrice { get; set; }
+        public decimal AskPrice { get; set; }
         [JsonProperty("openPrice")]
-        public double OpenPrice { get; set; }
+        public decimal OpenPrice { get; set; }
         [JsonProperty("highPrice")]
-        public double HighPrice { get; set; }
+        public decimal HighPrice { get; set; }
         [JsonProperty("lowPrice")]
-        public double LowPrice { get; set; }
+        public decimal LowPrice { get; set; }
         [JsonProperty("volume")]
-        public double Volume { get; set; }
+        public decimal Volume { get; set; }
         [JsonProperty("quoteVolume")]
-        public double QuoteVolume { get; set; }
+        public decimal QuoteVolume { get; set; }
         [JsonProperty("openTime")]
         public long OpenTime { get; set; }
         [JsonProperty("closeTime")]
