+++ Binance-Spot-API/Model/Order/Fill.cs
@@ -5,11 +5,11 @@ namespace Binance_Spot_API.Model.Order
     public class Fill
     {
         [JsonProperty("price")]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         [JsonProperty("qty")]
-        public double Quantity { get; set; }
+        public decimal Quantity { get; set; }
         [JsonProperty("commission")]
-        public double Commission { get; set; }
+        public decimal Commission { get; set; }
         [JsonProperty("commissionAsset")]
         public string CommissionAsset { get; set; }
